Fix admin web API bootstrap and transaction backups
