+++ app/webapi/routes/transactions.py
@@ -6,10 +6,11 @@ from typing import Any, Optional
 import logging
 
 from fastapi import APIRouter, Depends, Query, Security
-from sqlalchemy import and_, func, select, text
+from sqlalchemy import and_, func, select
 from sqlalchemy.ext.asyncio import AsyncSession
 
 from app.database.models import Transaction
+from app.database.utils import ensure_transaction_columns
 
 from ..dependencies import get_db_session, require_api_token
 from ..schemas.transactions import TransactionListResponse, TransactionResponse
@@ -17,75 +18,6 @@ from ..schemas.transactions import TransactionListResponse, TransactionResponse
 router = APIRouter()
 
 logger = logging.getLogger(__name__)
-
-_TRANSACTION_COLUMNS_READY = False
-
-
-async def _ensure_transaction_columns(db: AsyncSession) -> None:
-    global _TRANSACTION_COLUMNS_READY
-    if _TRANSACTION_COLUMNS_READY:
-        return
-
-    conn = await db.connection()
-    dialect = conn.dialect.name
-
-    async def column_exists(column: str) -> bool:
-        if dialect == "postgresql":
-            result = await conn.scalar(
-                text(
-                    """
-                    SELECT 1
-                    FROM information_schema.columns
-                    WHERE table_name = 'transactions' AND column_name = :column
-                    LIMIT 1
-                    """
-                ),
-                {"column": column},
-            )
-            return result is not None
-        if dialect == "sqlite":
-            result = await conn.execute(text("PRAGMA table_info('transactions')"))
-            return any(row[1] == column for row in result)
-        if dialect == "mysql":
-            result = await conn.scalar(
-                text(
-                    """
-                    SELECT 1
-                    FROM information_schema.columns
-                    WHERE table_schema = DATABASE()
-                      AND table_name = 'transactions'
-                      AND column_name = :column
-                    LIMIT 1
-                    """
-                ),
-                {"column": column},
-            )
-            return result is not None
-
-        logger.warning("Неизвестный драйвер БД %s для проверки колонок transactions", dialect)
-        return True
-
-    added = False
-
-    async def ensure_column(column: str, ddl_type: str) -> None:
-        nonlocal added
-        if await column_exists(column):
-            return
-        logger.info("Добавляем колонку transactions.%s", column)
-        await conn.execute(text(f"ALTER TABLE transactions ADD COLUMN {column} {ddl_type}"))
-        added = True
-
-    try:
-        await ensure_column("status", "VARCHAR(50)")
-        await ensure_column("currency", "VARCHAR(10)")
-        if added:
-            await db.commit()
-    except Exception as error:
-        await db.rollback()
-        logger.error("Не удалось обновить таблицу transactions: %s", error)
-        raise
-    else:
-        _TRANSACTION_COLUMNS_READY = True
 
 
 def _serialize(transaction: Transaction) -> TransactionResponse:
@@ -123,7 +55,7 @@ async def list_transactions(
     amount_max: Optional[int] = Query(default=None),
     currency: Optional[str] = Query(default=None),
 ) -> TransactionListResponse:
-    await _ensure_transaction_columns(db)
+    await ensure_transaction_columns(db)
 
     base_query = select(Transaction)
     conditions = []
